requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.9.0
argcomplete==1.10.3
astunparse==1.6.3
beautifulsoup4==4.8.2
blis==1.3.0
bs4==0.0.2
catalogue==2.0.10
certifi==2025.7.14
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
colorama==0.4.6
compressed-rtf==1.0.7
confection==0.1.5
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
cymem==2.0.11
deskew==1.5.3
distro==1.9.0
docx2txt==0.9
easyocr==1.7.2
ebcdic==1.1.1
et_xmlfile==2.0.0
fastapi==0.116.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.59.0
fsspec==2025.7.0
gast==0.6.0
geotext==0.4.0
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.73.1
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
imageio==2.37.0
IMAPClient==2.1.0
imutils==0.5.4
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
lxml==6.0.0
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.3.2
mpmath==1.3.0
murmurhash==1.0.13
namex==0.1.0
networkx==3.5
ninja==1.11.1.4
nltk==3.9.1
numpy==1.26.4
olefile==0.47
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.1
pdf2image==1.17.0
pdfminer.six==20191110
pillow==11.3.0
pip-licenses==5.0.0
pipdeptree==2.28.0
preshed==3.0.10
prettytable==3.16.0
protobuf==4.25.1
psutil==7.0.0
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-multipart==0.0.20
python-pptx==0.6.23
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.16.0
sentencepiece==0.2.0
shapely==2.1.1
shellingham==1.5.4
six==1.12.0
smart_open==7.3.0.post1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.8.1
SQLAlchemy==2.0.41
srsly==2.5.1
starlette==0.47.2
sympy==1.14.0
termcolor==3.1.0
thinc==8.3.6
tifffile==2025.6.11
tiktoken==0.9.0
tokenizers==0.21.2
tomli==2.2.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
xlrd==1.2.0
xlsxwriter==3.2.5
xmltodict==0.14.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
passlib[bcrypt]==1.7.4
pydantic-settings>=2.0.0
ollama>=0.1.0
